Optimizing Range Query fuction for OH

+++ frameworks/libs/distributeddb/gaussdb_rd/src/executor/document/grd_document_api.cpp
@@ -30,6 +30,10 @@ GRD_API int32_t GRD_CreateCollection(GRD_DB *db, const char *collectionName, con
     if (GRD_DocApiInfo.CreateCollectionApi == nullptr) {
         GRD_DocApiInfo = GetApiInfoInstance();
     }
+    if (GRD_DocApiInfo.CreateCollectionApi == nullptr) {
+        GLOGE("Fail to dlysm RD api symbol");
+        return GRD_INNER_ERR;
+    }
     return GRD_DocApiInfo.CreateCollectionApi(db, collectionName, optionStr, flags);
 }
 
@@ -37,6 +41,10 @@ GRD_API int32_t GRD_DropCollection(GRD_DB *db, const char *collectionName, uint3
 {
     if (GRD_DocApiInfo.DropCollectionApi == nullptr) {
         GRD_DocApiInfo = GetApiInfoInstance();
+    }
+    if (GRD_DocApiInfo.DropCollectionApi == nullptr) {
+        GLOGE("Fail to dlysm RD api symbol");
+        return GRD_INNER_ERR;
     }
     return GRD_DocApiInfo.DropCollectionApi(db, collectionName, flags);
 }
@@ -47,6 +55,10 @@ GRD_API int32_t GRD_UpdateDoc(GRD_DB *db, const char *collectionName, const char
     if (GRD_DocApiInfo.UpdateDocApi == nullptr) {
         GRD_DocApiInfo = GetApiInfoInstance();
     }
+    if (GRD_DocApiInfo.UpdateDocApi == nullptr) {
+        GLOGE("Fail to dlysm RD api symbol");
+        return GRD_INNER_ERR;
+    }
     return GRD_DocApiInfo.UpdateDocApi(db, collectionName, filter, update, flags);
 }
 
@@ -56,6 +68,10 @@ GRD_API int32_t GRD_UpsertDoc(GRD_DB *db, const char *collectionName, const char
     if (GRD_DocApiInfo.UpsertDocApi == nullptr) {
         GRD_DocApiInfo = GetApiInfoInstance();
     }
+    if (GRD_DocApiInfo.UpsertDocApi == nullptr) {
+        GLOGE("Fail to dlysm RD api symbol");
+        return GRD_INNER_ERR;
+    }
     return GRD_DocApiInfo.UpsertDocApi(db, collectionName, filter, document, flags);
 }
 
@@ -63,6 +79,10 @@ GRD_API int32_t GRD_InsertDoc(GRD_DB *db, const char *collectionName, const char
 {
     if (GRD_DocApiInfo.InsertDocApi == nullptr) {
         GRD_DocApiInfo = GetApiInfoInstance();
+    }
+    if (GRD_DocApiInfo.InsertDocApi == nullptr) {
+        GLOGE("Fail to dlysm RD api symbol");
+        return GRD_INNER_ERR;
     }
     return GRD_DocApiInfo.InsertDocApi(db, collectionName, document, flags);
 }
@@ -72,6 +92,10 @@ GRD_API int32_t GRD_DeleteDoc(GRD_DB *db, const char *collectionName, const char
     if (GRD_DocApiInfo.DeleteDocApi == nullptr) {
         GRD_DocApiInfo = GetApiInfoInstance();
     }
+    if (GRD_DocApiInfo.DeleteDocApi == nullptr) {
+        GLOGE("Fail to dlysm RD api symbol");
+        return GRD_INNER_ERR;
+    }
     return GRD_DocApiInfo.DeleteDocApi(db, collectionName, filter, flags);
 }
 
@@ -80,6 +104,10 @@ GRD_API int32_t GRD_FindDoc(GRD_DB *db, const char *collectionName, Query query,
 {
     if (GRD_DocApiInfo.FindDocApi == nullptr) {
         GRD_DocApiInfo = GetApiInfoInstance();
+    }
+    if (GRD_DocApiInfo.FindDocApi == nullptr) {
+        GLOGE("Fail to dlysm RD api symbol");
+        return GRD_INNER_ERR;
     }
     return GRD_DocApiInfo.FindDocApi(db, collectionName, query, flags, resultSet);
 }
